requirements.txt
Django==6.0.3
celery==5.6.2
psycopg2-binary==2.9.11
pillow==12.1.1
pyusb==1.3.1
attrs==25.3.0
structlog==25.5.0
envparse==0.2.0
python-dotenv==1.1.0
qrcode[pil]==8.1
piexif==1.1.3
